+++ script/MangroveDeployer.s.sol
@@ -6,7 +6,9 @@ import "mgv_src/periphery/MgvReader.sol";
 import {MgvCleaner} from "mgv_src/periphery/MgvCleaner.sol";
 import {MgvOracle} from "mgv_src/periphery/MgvOracle.sol";
 import {IMangrove} from "mgv_src/IMangrove.sol";
-import {Deployer} from "./Deployer.sol";
+import {Deployer} from "./lib/Deployer.sol";
+import {MgvCleanerDeployer} from "./periphery/MgvCleaner.s.sol";
+import {MgvReaderDeployer} from "./periphery/MgvReader.s.sol";
 
 contract MangroveDeployer is Deployer {
   Mangrove public mgv;
@@ -25,19 +27,25 @@ contract MangroveDeployer is Deployer {
 
   function innerRun(address chief, uint gasprice, uint gasmax) public {
     broadcast();
-    mgv = new Mangrove({governance: chief, gasprice: gasprice, gasmax: gasmax});
+    if (forMultisig) {
+      mgv = new Mangrove{salt:salt}({governance: chief, gasprice: gasprice, gasmax: gasmax});
+    } else {
+      mgv = new Mangrove({governance: chief, gasprice: gasprice, gasmax: gasmax});
+    }
     fork.set("Mangrove", address(mgv));
 
-    broadcast();
-    reader = new MgvReader({mgv: payable(mgv)});
-    fork.set("MgvReader", address(reader));
+    (new MgvReaderDeployer()).innerRun(address(mgv));
+    reader = MgvReader(fork.get("MgvReader"));
+
+    (new MgvCleanerDeployer()).innerRun(address(mgv));
+    cleaner = MgvCleaner(fork.get("MgvCleaner"));
 
     broadcast();
-    cleaner = new MgvCleaner({mgv: address(mgv)});
-    fork.set("MgvCleaner", address(cleaner));
-
-    broadcast();
-    oracle = new MgvOracle({governance_: chief, initialMutator_: chief});
+    if (forMultisig) {
+      oracle = new MgvOracle{salt: salt}({governance_: chief, initialMutator_: chief});
+    } else {
+      oracle = new MgvOracle({governance_: chief, initialMutator_: chief});
+    }
     fork.set("MgvOracle", address(oracle));
   }
 }